Wave forms initial commit
This project is intended to make generating pretty timing diagrams
trivial. They will be entered using a simple, and readable plain text
format. The package will render it into an SVG with appropiate options.

Currently, the project has the start of the waveform package. It stores
each waveform in a object, contained as a list in a parent.

Signed-off-by: Jacob von Chorus <jacobvonchorus@cwphoto.ca>