Produces waveform svgs
Using spacings laid out globally in the file (tuning knobs), the svg
waveforms are generated from the input text.

Each line is one lineheight in height. Text is anchored to the bottom of
each line. The ticks appear at either the very top or bottom of the
line.

There is a 'spacing' between each line, typically less than a line
height. This spacing is also before the first line. A spacing was added
between the (right-justified) text and the waves.

A spacing was also added below and to the right of the waves to help
viewers ensure their display program isn't truncating anything.

Signed-off-by: Jacob von Chorus <jacobvonchorus@cwphoto.ca>